chore: Organizando codigo (Movido PieceExtension para um file .cs)
