+++ LagChessApplication/Domains/Player.cs
@@ -1,5 +1,5 @@
-﻿using LagChessApplication.Domains.Enums;
-using LagChessApplication.Domains.Pieces;
+﻿using LagChessApplication.Domains.Pieces;
+using LagChessApplication.Extensions;
 using LagChessApplication.Interfaces;
 
 namespace LagChessApplication.Domains
@@ -28,23 +28,23 @@ namespace LagChessApplication.Domains
                 Name = name,
                 Pieces =
                 [
-                    PieceBase.CreatePieceWhite<Rook>(1, 1),
-                    PieceBase.CreatePieceWhite<Knight>(2, 1),
-                    PieceBase.CreatePieceWhite<Bishop>(3, 1),
-                    PieceBase.CreatePieceWhite<Queen>(4, 1),
-                    PieceBase.CreatePieceWhite<King>(5, 1),
-                    PieceBase.CreatePieceWhite<Bishop>(6, 1),
-                    PieceBase.CreatePieceWhite<Knight>(7, 1),
-                    PieceBase.CreatePieceWhite<Rook>(8, 1),
+                    PieceExtension.CreatePieceWhite<Rook>(1, 1),
+                    PieceExtension.CreatePieceWhite<Knight>(2, 1),
+                    PieceExtension.CreatePieceWhite<Bishop>(3, 1),
+                    PieceExtension.CreatePieceWhite<Queen>(4, 1),
+                    PieceExtension.CreatePieceWhite<King>(5, 1),
+                    PieceExtension.CreatePieceWhite<Bishop>(6, 1),
+                    PieceExtension.CreatePieceWhite<Knight>(7, 1),
+                    PieceExtension.CreatePieceWhite<Rook>(8, 1),
 
-                    PieceBase.CreatePieceWhite<Pawn>(1, 2),
-                    PieceBase.CreatePieceWhite<Pawn>(2, 2),
-                    PieceBase.CreatePieceWhite<Pawn>(3, 2),
-                    PieceBase.CreatePieceWhite<Pawn>(4, 2),
-                    PieceBase.CreatePieceWhite<Pawn>(5, 2),
-                    PieceBase.CreatePieceWhite<Pawn>(6, 2),
-                    PieceBase.CreatePieceWhite<Pawn>(7, 2),
-                    PieceBase.CreatePieceWhite<Pawn>(8, 2)
+                    PieceExtension.CreatePieceWhite<Pawn>(1, 2),
+                    PieceExtension.CreatePieceWhite<Pawn>(2, 2),
+                    PieceExtension.CreatePieceWhite<Pawn>(3, 2),
+                    PieceExtension.CreatePieceWhite<Pawn>(4, 2),
+                    PieceExtension.CreatePieceWhite<Pawn>(5, 2),
+                    PieceExtension.CreatePieceWhite<Pawn>(6, 2),
+                    PieceExtension.CreatePieceWhite<Pawn>(7, 2),
+                    PieceExtension.CreatePieceWhite<Pawn>(8, 2)
                 ]
             };
         }
@@ -55,23 +55,23 @@ namespace LagChessApplication.Domains
                 Name = name,
                 Pieces =
                 [
-                    PieceBase.CreatePieceBlack<Rook>(1, 8),
-                    PieceBase.CreatePieceBlack<Knight>(2, 8),
-                    PieceBase.CreatePieceBlack<Bishop>(3, 8),
-                    PieceBase.CreatePieceBlack<Queen>(4, 8),
-                    PieceBase.CreatePieceBlack<King>(5, 8),
-                    PieceBase.CreatePieceBlack<Bishop>(6, 8),
-                    PieceBase.CreatePieceBlack<Knight>(7, 8),
-                    PieceBase.CreatePieceBlack<Rook>(8, 8),
+                    PieceExtension.CreatePieceBlack<Rook>(1, 8),
+                    PieceExtension.CreatePieceBlack<Knight>(2, 8),
+                    PieceExtension.CreatePieceBlack<Bishop>(3, 8),
+                    PieceExtension.CreatePieceBlack<Queen>(4, 8),
+                    PieceExtension.CreatePieceBlack<King>(5, 8),
+                    PieceExtension.CreatePieceBlack<Bishop>(6, 8),
+                    PieceExtension.CreatePieceBlack<Knight>(7, 8),
+                    PieceExtension.CreatePieceBlack<Rook>(8, 8),
 
-                    PieceBase.CreatePieceBlack<Pawn>(1, 7),
-                    PieceBase.CreatePieceBlack<Pawn>(2, 7),
-                    PieceBase.CreatePieceBlack<Pawn>(3, 7),
-                    PieceBase.CreatePieceBlack<Pawn>(4, 7),
-                    PieceBase.CreatePieceBlack<Pawn>(5, 7),
-                    PieceBase.CreatePieceBlack<Pawn>(6, 7),
-                    PieceBase.CreatePieceBlack<Pawn>(7, 7),
-                    PieceBase.CreatePieceBlack<Pawn>(8, 7)
+                    PieceExtension.CreatePieceBlack<Pawn>(1, 7),
+                    PieceExtension.CreatePieceBlack<Pawn>(2, 7),
+                    PieceExtension.CreatePieceBlack<Pawn>(3, 7),
+                    PieceExtension.CreatePieceBlack<Pawn>(4, 7),
+                    PieceExtension.CreatePieceBlack<Pawn>(5, 7),
+                    PieceExtension.CreatePieceBlack<Pawn>(6, 7),
+                    PieceExtension.CreatePieceBlack<Pawn>(7, 7),
+                    PieceExtension.CreatePieceBlack<Pawn>(8, 7)
                 ]
             };
         }
